aws/internal/tfresource/resource.go
package tfresource

import (
	"context"
	"errors"

	multierror "github.com/hashicorp/go-multierror"
	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
)

// Delete calls the specified resource's delete handler.
func Delete(resource *schema.Resource, d *schema.ResourceData, meta interface{}) error {
	if resource.DeleteContext != nil || resource.DeleteWithoutTimeout != nil {
		var diags diag.Diagnostics
		var err *multierror.Error

		if resource.DeleteContext != nil {
			diags = resource.DeleteContext(context.Background(), d, meta)
		} else {
			diags = resource.DeleteWithoutTimeout(context.Background(), d, meta)
		}

		for i := range diags {
			if diags[i].Severity == diag.Error {
				err = multierror.Append(err, errors.New(diags[i].Summary))
			}
		}

		return err.ErrorOrNil()
	}

	return resource.Delete(d, meta)
}
